fix(cli): honour the global --principal flag in astrid init for per-principal provisioning
astrid init hardcoded PrincipalId::default() for the Distro.lock path, per-capsule env files, and the install read-back, so `astrid init --principal <x>` provisioned `default`'s home and left a scoped principal with an empty capsule view. The capsule FILES already installed under the resolved principal (via the *_for_principal install lib), so the lock recorded empty hashes and onboarding read the wrong manifest.

Thread the process-wide principal (crate::principal::current(), the same source `astrid capsule install` reads) through run_init and the offline shuttle path: lock path, write_env_files, onboard_llm_providers, and every resolve_target_dir read-back now scope to the resolved principal.

Provisioning honesty: do not print success or write Distro.lock when every install failed (bail non-zero — a fresh lock would wedge re-runs); a partial provision writes the lock and reports the true N/total count. Make the shell-PATH edit idempotent via a tested pure predicate. Authenticate GitHub API release/source resolution with GH_TOKEN/GITHUB_TOKEN when present (5000/hr vs anonymous 60/hr), degrading to anonymous when absent; reqwest strips the token on cross-host redirects.

Rate limit (part 3): assessed — astrid init does NOT trigger the kernel's 5/min ReloadCapsule throttle. Batch install (install_capsule_batch) deliberately skips the live-reload nudge; capsules load on next daemon start. No init change needed; documented as a kernel-side note.

Regression test write_env_files_targets_the_given_principal fails without the fix (env lands under `default`). Unit tests added for should_write_lock and rc_configures_path. Extracted init.rs tests to init_tests.rs to stay under the 1000-line file cap.

Closes #1095
